refactor(exception-catch): Most of update and create return bool for properly identify error

+++ backend/services/auth_service.py
@@ -9,10 +9,11 @@ class AuthService:
         self.session_repo = session_repo
 
     def register(self, username, password, first, last, role) -> bool:
+        """Return None when username exist, False when role not allowed, True otherwise"""
         if role not in AVAILABLE_ROLE:
             return False
         if self.user_repo.is_username_exist(username):
-            return False
+            return None
         self.user_repo.create_user(username, password, first, last, role)
         return True
 
@@ -22,14 +23,24 @@ class AuthService:
             return None
         return self.session_repo.create_session(user.id)
 
-    def get_session_by_id(self, session_id):
-        return self.session_repo.get_active_session_by_id(session_id)
+    def get_session_by_id(self, session_id, is_active=True):
+        return self.session_repo.get_session_by_id(session_id, is_active)
 
     def refresh_session(self, session_id):
+        """Return None when session_id not found otherwise True"""
+        session = self.get_session_by_id(session_id)
+        if session is None:
+            return None
         self.session_repo.refresh_session(session_id)
+        return True
 
     def logout(self, session_id):
+        """Return None when session_id not found otherwise True"""
+        session = self.get_session_by_id(session_id, False)
+        if session is None:
+            return None
         self.session_repo.delete_session(session_id)
+        return True
 
     def get_user_by_id(self, user_id):
         return self.user_repo.get_user_by_id(user_id)
@@ -40,7 +51,11 @@ class AuthService:
     def change_password(self, username, old_password, new_password) -> bool:
         return self.user_repo.change_password(username, old_password, new_password)
 
-    def update_profile(self, user_id, username, first, last) -> bool:
+    def update_profile(self, user_id, username, first, last):
+        """Return None when user_id not found, False when username exist, True otherwise"""
+        user = self.get_user_by_id(user_id)
+        if user is None:
+            return None
         if self.user_repo.is_username_exist(username):
             return False
         self.user_repo.update_user(user_id, username, first, last)
